requirements.txt
aws_cdk.aws_s3==1.152
aws_cdk.aws_kinesisfirehose==1.152
aws_cdk.aws_rds==1.152
aws_cdk.aws_dms==1.152
aws_cdk.aws_glue==1.152
aws_cdk.aws_athena==1.152
aws_cdk.aws_mwaa==1.152
aws_cdk.aws_s3_deployment==1.152
aws_cdk.aws_redshift==1.152
fake_web_events
boto3
psycopg2-binary
requests
backoff
ratelimit